Fix signed/unsigned comparison warning in schedulers

+++ simulatorCore/src/cppScheduler.cpp
@@ -102,7 +102,7 @@ void *CPPScheduler::startPaused(/*void *param*/) {
 					if (!eventsMap.empty()) {
 						first=eventsMap.begin();
 						pev = (*first).second;
-						while (!eventsMap.empty() && pev->date <= chrono::duration_cast<us>(systemCurrentTimeMax).count()) {
+						while (!eventsMap.empty() && pev->date <= static_cast<uint64_t>(chrono::duration_cast<us>(systemCurrentTimeMax).count())) {
 							first=eventsMap.begin();
 							pev = (*first).second;
 							currentDate = pev->date;
